+++ CHANGELOG.md
@@ -6,3 +6,10 @@
 - Add tips.csv, file to analyze
 - Add Python file where program will be run
 
+## 2024-10-29
+### Added
+- Implement Spark session
+- Implement read file
+- Implement data cleaning and transformation
+- Implement train and test data
+- Implement model